Created special `TileGraphics` object for handling graphics
